Update Coordinate.java
fixing setters to use doubles, not ints

+++ src/main/java/ca/dhlevi/libnoise/spatial/Coordinate.java
@@ -20,7 +20,7 @@ public class Coordinate
         return x;
     }
 
-    public void setX(int x)
+    public void setX(double x)
     {
         this.x = x;
     }
@@ -30,7 +30,7 @@ public class Coordinate
         return y;
     }
 
-    public void setY(int y)
+    public void setY(double y)
     {
         this.y = y;
     }
